requirements.txt
torch>=2.1
einops
numpy
scipy
matplotlib
datasets
tokenizers
transformers
composer
biopython
pillow